pyproject.toml
[tool.poetry]
name = "djangoback"
version = "0.1.0"
description = ""
authors = ["sds7629 <sds7629@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.2"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

